backend/Healthconnect/requirements.txt
asgiref==3.6.0
Django==3.2.15
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.2.0
PyJWT==2.6.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.3
